+++ package.json
@@ -2,7 +2,7 @@
   "name": "react-tabs",
   "version": "0.8.2",
   "description": "React tabs component",
-  "main": "lib/main.js",
+  "main": "lib/index.js",
   "scripts": {
     "clean": "rimraf lib",
     "build:commonjs": "babel src/ --out-dir lib/ --ignore __tests__,__mocks__",
