+++ src/__tests__/index-test.js
@@ -1,12 +1,11 @@
 /* eslint-env jest */
-/* eslint-disable import/no-named-as-default-member */
-import { Tab, Tabs, TabList, TabPanel } from '../main';
+import { Tab, Tabs, TabList, TabPanel } from '../index';
 import TabComponent from '../components/Tab';
 import TabListComponent from '../components/TabList';
 import TabsComponent from '../components/Tabs';
 import TabPanelComponent from '../components/TabPanel';
 
-describe('main.js', () => {
+describe('index.js', () => {
   it('should correctly export all component as named export', () => {
     expect(Tab).toEqual(TabComponent);
     expect(TabList).toEqual(TabListComponent);
